Reorganise player components and composables under players/ subtrees
Group all preview / playlist / shared-playlist player code (annotations,
bars, viewers, progress, sides, players, headers) under
src/components/players/<area>/ with dedicated subfolders, and move the
player-specific composables (annotation, annotationBroadcast,
comparison, playlistComparison, previewShortcuts, sharedAnnotation)
under src/composables/players/. PlaylistPlayer.vue and PlaylistedEntity
move there too, leaving src/components/pages/playlists/ empty.

All imports updated. CLAUDE.md adjusted to reflect the new layout.
