+++ src/components/pages/Edit.vue
@@ -231,8 +231,8 @@ import EntityThumbnail from '@/components/widgets/EntityThumbnail.vue'
 import EntityTimeLogs from '@/components/pages/entities/EntityTimeLogs.vue'
 import MetadataValue from '@/components/widgets/MetadataValue.vue'
 import PageSubtitle from '@/components/widgets/PageSubtitle.vue'
-import PreviewPlayer from '@/components/previews/PreviewPlayer.vue'
-import PreviewsPerTaskType from '@/components/previews/PreviewsPerTaskType.vue'
+import PreviewPlayer from '@/components/players/players/PreviewPlayer.vue'
+import PreviewsPerTaskType from '@/components/players/bars/PreviewsPerTaskType.vue'
 import RouteSectionTabs from '@/components/widgets/RouteSectionTabs.vue'
 import Schedule from '@/components/widgets/Schedule.vue'
 /* eslint-enable no-unused-vars */
