+++ src/components/pages/Playlist.vue
@@ -615,7 +615,7 @@ import EditPlaylistModal from '@/components/modals/EditPlaylistModal.vue'
 import ErrorText from '@/components/widgets/ErrorText.vue'
 import LightEntityThumbnail from '@/components/widgets/LightEntityThumbnail.vue'
 import PageSubtitle from '@/components/widgets/PageSubtitle.vue'
-import PlaylistPlayer from '@/components/pages/playlists/PlaylistPlayer.vue'
+import PlaylistPlayer from '@/components/players/players/PlaylistPlayer.vue'
 import SearchField from '@/components/widgets/SearchField.vue'
 import Spinner from '@/components/widgets/Spinner.vue'
 import TaskTypeName from '@/components/widgets/TaskTypeName.vue'
