+++ src/components/players/bars/BrowsingBar.vue
