+++ src/components/players/bars/PlayerAnnotationBar.vue
