+++ src/components/players/bars/PreviewsPerTaskType.vue
