+++ src/components/players/headers/SharedPlaylistHeader.vue
