+++ src/components/players/players/PlaylistPlayer.vue
@@ -870,12 +870,12 @@ import { useI18n } from 'vue-i18n'
 import { useStore } from 'vuex'
 import WaveSurfer from 'wavesurfer.js'
 
-import { useAnnotation } from '@/composables/annotation'
-import { useAnnotationBroadcast } from '@/composables/annotationBroadcast'
+import { useAnnotation } from '@/composables/players/annotation'
+import { useAnnotationBroadcast } from '@/composables/players/annotationBroadcast'
 import { useFullScreen } from '@/composables/fullScreen'
-import { usePlaylistComparison } from '@/composables/playlistComparison'
+import { usePlaylistComparison } from '@/composables/players/playlistComparison'
 import { usePreviewRoom } from '@/composables/previewRoom'
-import { usePreviewShortcuts } from '@/composables/previewShortcuts'
+import { usePreviewShortcuts } from '@/composables/players/previewShortcuts'
 import preferences from '@/lib/preferences'
 import {
   isModelPreview,
@@ -897,21 +897,21 @@ import NotifyClientModal from '@/components/modals/NotifyClientModal.vue'
 import SelectTaskTypeModal from '@/components/modals/SelectTaskTypeModal.vue'
 import SharePlaylistModal from '@/components/modals/SharePlaylistModal.vue'
 
-import PlaylistedEntity from '@/components/pages/playlists/PlaylistedEntity.vue'
+import PlaylistedEntity from '@/components/players/players/PlaylistedEntity.vue'
 
-import AnnotationCanvas from '@/components/previews/AnnotationCanvas.vue'
-import MultiPictureViewer from '@/components/previews/MultiPictureViewer.vue'
-import MultiVideoViewer from '@/components/previews/MultiVideoViewer.vue'
-import ObjectViewer from '@/components/previews/ObjectViewer.vue'
-import PdfViewer from '@/components/previews/PdfViewer.vue'
-import PictureViewer from '@/components/previews/PictureViewer.vue'
-import PlayerAnnotationBar from '@/components/previews/PlayerAnnotationBar.vue'
-import PlayerComparisonBar from '@/components/previews/PlayerComparisonBar.vue'
-import PlayerPlaybackBar from '@/components/previews/PlayerPlaybackBar.vue'
+import AnnotationCanvas from '@/components/players/annotations/AnnotationCanvas.vue'
+import MultiPictureViewer from '@/components/players/viewers/MultiPictureViewer.vue'
+import MultiVideoViewer from '@/components/players/viewers/MultiVideoViewer.vue'
+import ObjectViewer from '@/components/players/viewers/ObjectViewer.vue'
+import PdfViewer from '@/components/players/viewers/PdfViewer.vue'
+import PictureViewer from '@/components/players/viewers/PictureViewer.vue'
+import PlayerAnnotationBar from '@/components/players/bars/PlayerAnnotationBar.vue'
+import PlayerComparisonBar from '@/components/players/bars/PlayerComparisonBar.vue'
+import PlayerPlaybackBar from '@/components/players/bars/PlayerPlaybackBar.vue'
 // eslint-disable-next-line no-unused-vars -- shadowed by setPlaylistProgress / playlistProgressRef in script
-import PlaylistProgress from '@/components/previews/PlaylistProgress.vue'
-import SoundViewer from '@/components/previews/SoundViewer.vue'
-import VideoProgress from '@/components/previews/VideoProgress.vue'
+import PlaylistProgress from '@/components/players/progress/PlaylistProgress.vue'
+import SoundViewer from '@/components/players/viewers/SoundViewer.vue'
+import VideoProgress from '@/components/players/progress/VideoProgress.vue'
 
 import ButtonSimple from '@/components/widgets/ButtonSimple.vue'
 // eslint-disable-next-line no-unused-vars -- shadowed by const previewRoom in script
