+++ src/components/players/players/SharedPlaylistPlayer.vue
@@ -223,16 +223,16 @@ import { usePanzoomSync } from '@/composables/panzoom'
 import { isMoviePreview, isPicturePreview, isSoundPreview } from '@/lib/preview'
 import { floorToFrame, formatTime } from '@/lib/video'
 
-import PictureViewer from '@/components/previews/PictureViewer.vue'
-import PlaylistedEntity from '@/components/pages/playlists/PlaylistedEntity.vue'
-import PlaylistProgress from '@/components/previews/PlaylistProgress.vue'
-import MultiVideoViewer from '@/components/previews/MultiVideoViewer.vue'
-import SharedAnnotationOverlay from '@/components/previews/SharedAnnotationOverlay.vue'
-import SharedCommentsPanel from '@/components/previews/SharedCommentsPanel.vue'
-import SharedPlaylistButtonBar from '@/components/previews/SharedPlaylistButtonBar.vue'
-import SharedPlaylistHeader from '@/components/previews/SharedPlaylistHeader.vue'
-import SoundViewer from '@/components/previews/SoundViewer.vue'
-import VideoProgress from '@/components/previews/VideoProgress.vue'
+import PictureViewer from '@/components/players/viewers/PictureViewer.vue'
+import PlaylistedEntity from '@/components/players/players/PlaylistedEntity.vue'
+import PlaylistProgress from '@/components/players/progress/PlaylistProgress.vue'
+import MultiVideoViewer from '@/components/players/viewers/MultiVideoViewer.vue'
+import SharedAnnotationOverlay from '@/components/players/annotations/SharedAnnotationOverlay.vue'
+import SharedCommentsPanel from '@/components/players/sides/SharedCommentsPanel.vue'
+import SharedPlaylistButtonBar from '@/components/players/bars/SharedPlaylistButtonBar.vue'
+import SharedPlaylistHeader from '@/components/players/headers/SharedPlaylistHeader.vue'
+import SoundViewer from '@/components/players/viewers/SoundViewer.vue'
+import VideoProgress from '@/components/players/progress/VideoProgress.vue'
 import Spinner from '@/components/widgets/Spinner.vue'
 
 // Props / Emits
