+++ src/components/players/progress/PlaylistProgress.vue
