+++ src/components/players/progress/VideoProgress.vue
