+++ src/components/players/viewers/ObjectViewer.vue
