+++ src/components/players/viewers/PictureViewer.vue
