+++ src/composables/players/previewShortcuts.js
